Explicitly disable FF
